Proper Transform Class
A more refined transform class. No longer requires an update to be
called each frame. Requesting the transform's world matrix always
returns the most up-to-date information. Parents don't know about
children and don't need to.

Setting an object's world position is absolute as it should be. Setting
it's local position is relative as it should be.

Added xml comments to the header files of Planet.h and Transform.h for
code clarity.