Clean Burn
Everything is now using unique_ptr when an object owns that dynamic
memory. It is deallocated automatically except when it's static in which
case I alllocate and deallocate it on init() and quit()

Files should not be reincluding vector, memory, and glm as this was put
into a master header which defines which classes to use along with quick
typedefs to use for smart pointers. Also a PI define was added.

Creating an object of MyApplication and then deleting it leaves no trace
on the heap from my code. There seems to be something left over from
openGL so it's possible that I'm not terminating it properly.

ImGui is somehow bugged as it cannot run twice in the same program
without getting an assert error. Not sure what's going on with that.
Program runs fine once, but if you delete it and then try to start it up
again ImGui assert fails saying delta time is not >= 0