Shader Variables
Began working on a grammar for putting shader variables into the
inspector automagically by using comments. Looks something like

// Inspector Color()
uniform vec4 MaterialColor = vec4(0, 0, 0, 1);

which will create a modifiable variable for the shader in the inspector
called "Material Color" of type vec4 and display it as a color type of
Gui element.